+++ spec/spec_helper.rb
@@ -1,5 +1,15 @@
 # frozen_string_literal: true
 
+require "simplecov"
+
+SimpleCov.start do
+  enable_coverage :branch
+  add_filter "/spec/"
+  track_files "lib/**/*.rb"
+  minimum_coverage line: 100, branch: 100
+end
+
+require "active_record"
 require "strong_schema"
 
 RSpec.configure do |config|